moving invaders
I had issues which required me to change all posX and posY into floats as rounding made certain speeds uncalculable due to int to float rounding errors

+++ Entity.cpp
@@ -1,6 +1,6 @@
 #include "Entity.h"
 
-Entity::Entity(Texture2D* texture, int posX, int posY)
+Entity::Entity(Texture2D* texture, float posX, float posY)
 {
 	this->texture = texture;
 	this->posX = posX;
